tests/wpt/web-platform-tests/client-hints/viewport-size/viewport-size-window-different-dimensions.https.html

<!DOCTYPE html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(t => {
  return new Promise(resolve => {
    window.addEventListener("message", t.step_func(e => {
      assert_equals(e.data.viewportWidth, "503");
      assert_equals(e.data.viewportHeight, "614");
      resolve();
    }));
  });
});
window.open("../resources/viewport-frame.py", "", "width=503,height=614");
</script>
