third_party/blink/web_tests/platform/linux-chrome/external/wpt/client-hints/viewport-size/viewport-size-window-different-dimensions.https-expected.txt

This is a testharness.js-based test.
[FAIL] viewport-size-window-different-dimensions
  assert_equals: expected "614" but got "600"
Harness: the test ran to completion.
